Things cocci missed, and lots of little prototypes.
This is the not so fun part. I'm done for the night :-)
But we're close!

+++ sys/src/libc/port/memset.c
@@ -11,7 +11,7 @@
 #include	<libc.h>
 
 void*
-memset(void *ap, int c, uint32_t n)
+memset(void *ap, int c, unsigned long n)
 {
 	char *p;
